threads ATOMIC_LOCK is working! Removed deadlock ts::con::prints <-> ts::mem32::alloc by using printf, created debug() print dedicated to be based on mallock or stack memory buffer so deadlock couldnt occur or debug could be redirected to log if neaded

+++ time/tsoft_time.h
@@ -6,8 +6,8 @@
 //---------------------------------------------------------------------------
 namespace ts { namespace time {
 //------------------------------------
-extern uint64_t		    	__stdcall	clock_us(void);
-extern uint64_t			    __stdcall	clock_ms(void);
+extern double		    	__stdcall	clock_us(void);
+extern double			    __stdcall	clock_ms(void);
 extern void 		        __stdcall	wait_us(const __int32 a_microseconds);
 extern void 		        __stdcall	wait_ms(const __int32 a_milliseconds);
